UPDATING DOCUMENTS
