+++ app/src/main/AndroidManifest.xml
@@ -13,38 +13,52 @@
         android:theme="@style/Theme.TaskApp"
         tools:targetApi="31">
         <activity
-            android:name=".Category"
+            android:name=".Vistas.Cuadernos.BooksMenu"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".Categorias"
+            android:name=".Vistas.Configuracion.Settings"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.HomeApp"
+            android:name=".Vistas.Categorias.CategoryList"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.SyncUp"
+            android:name=".Vistas.Categorias.Addcategory"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".activities.CreateAccount"
-            android:parentActivityName=".activities.LogIn" />
+            android:name=".Vistas.Inicio.HomeApp"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
         <activity
-            android:name=".activities.Pantalla_carga"
+            android:name=".Vistas.Precarga.SyncUp"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
+            android:name=".Vistas.Autenticacion.CreateAccount"
+            android:parentActivityName=".Vistas.Autenticacion.LogIn" />
+        <activity
+            android:name=".Vistas.Precarga.ChargingScreen"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -53,10 +67,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".Vistas.Inicio.MainActivity"
             android:exported="true" />
         <activity
-            android:name=".activities.LogIn"
+            android:name=".Vistas.Autenticacion.LogIn"
             android:exported="true"
             android:label="@string/title_activity_login" />
     </application>
