+++ app/src/main/java/com/example/task_app/Controladores/Almacenamiento/User.java
@@ -1,4 +1,4 @@
-package com.example.task_app.model.firebase_authentication;
+package com.example.task_app.Controladores.Almacenamiento;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
