+++ app/src/main/java/com/example/task_app/Modelos/Nota.java
@@ -1,4 +1,4 @@
-package com.example.task_app.Class;
+package com.example.task_app.Modelos;
 
 import java.util.Comparator;
 import java.util.Date;
