+++ app/src/main/java/com/example/task_app/Modelos/Usuario.java
@@ -1,4 +1,4 @@
-package com.example.task_app.Class;
+package com.example.task_app.Modelos;
 
 import android.media.Image;
