+++ app/src/main/java/com/example/task_app/Modelos/na_AdapterNotas.java
@@ -1,4 +1,4 @@
-package com.example.task_app.Class;
+package com.example.task_app.Modelos;
 
 import android.view.LayoutInflater;
 import android.view.View;
