+++ app/src/main/java/com/example/task_app/Vistas/Precarga/ChargingScreen.java
@@ -1,4 +1,4 @@
-package com.example.task_app.activities;
+package com.example.task_app.Vistas.Precarga;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -7,6 +7,7 @@ import android.os.Bundle;
 import android.os.Handler;
 
 import com.example.task_app.R;
+import com.example.task_app.Vistas.Inicio.MainActivity;
 
 public class ChargingScreen extends AppCompatActivity {
