+++ app/src/main/java/com/example/task_app/Vistas/Tareas/AllWorksActivity.java
@@ -1,11 +1,9 @@
-package com.example.task_app.task;
+package com.example.task_app.Vistas.Tareas;
 
 import android.content.Intent;
-import android.media.Image;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ArrayAdapter;
-import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.ListView;
