+++ app/src/main/java/com/example/task_app/Vistas/Tareas/NewWorkActivity.java
@@ -1,14 +1,12 @@
-package com.example.task_app.task;
+package com.example.task_app.Vistas.Tareas;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
-import android.widget.Toast;
 
 import com.example.task_app.R;
