+++ app/src/main/res/layout/activity_all_works.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".task.AllWorksActivity">
+    tools:context=".Vistas.Tareas.AllWorksActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
