+++ app/src/main/res/layout/activity_books_menu.xml
@@ -9,22 +9,26 @@
         android:id="@+id/back_button"
         android:layout_width="50dp"
         android:layout_height="39dp"
+        android:layout_marginStart="16dp"
+        android:layout_marginTop="16dp"
         android:background="@drawable/boton_semiredondo"
-        android:text="<"
+        android:text="BACK"
         android:textSize="20dp"
-        tools:layout_editor_absoluteX="16dp"
-        tools:layout_editor_absoluteY="16dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
         tools:ignore="MissingConstraints" />
 
     <Button
         android:id="@+id/op_button"
         android:layout_width="50dp"
         android:layout_height="39dp"
+        android:layout_marginTop="16dp"
+        android:layout_marginEnd="16dp"
         android:background="@drawable/boton_semiredondo"
         android:text="☰"
         android:textSize="20dp"
-        tools:layout_editor_absoluteX="345dp"
-        tools:layout_editor_absoluteY="16dp" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <LinearLayout
         android:id="@+id/linearLayout2"
@@ -48,14 +52,16 @@
         style="@style/Widget.AppCompat.Button"
         android:layout_width="65dp"
         android:layout_height="65dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="16dp"
         android:background="@drawable/boton_redondo"
         android:outlineAmbientShadowColor="#FFFFFF"
         android:shadowColor="#FBFBFB"
         android:text="+"
         android:textAlignment="center"
         android:textSize="34sp"
-        tools:layout_editor_absoluteX="312dp"
-        tools:layout_editor_absoluteY="650dp" />
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
 
     <TextView
         android:id="@+id/an_cuadernosmenu"
