+++ app/src/main/res/layout/activity_charging_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.ChargingScreen">
+    tools:context=".Vistas.Precarga.ChargingScreen">
 
     <TextView
         android:id="@+id/textView7"
