+++ app/src/main/res/layout/activity_create_account.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.CreateAccount">
+    tools:context=".Vistas.Autenticacion.CreateAccount">
 
     <TextView
         android:id="@+id/textView17"
