+++ app/src/main/res/layout/activity_home_app.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.HomeApp">
+    tools:context=".Vistas.Inicio.HomeApp">
 
     <TextView
         android:id="@+id/textView8"
@@ -108,7 +108,7 @@
                         app:layout_constraintTop_toTopOf="parent" />
 
                     <Button
-                        android:id="@+id/button2"
+                        android:id="@+id/Home_Categorias"
                         android:layout_width="0dp"
                         android:layout_height="70dp"
                         android:layout_marginStart="8dp"
