+++ app/src/main/res/layout/activity_new_work.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".task.NewWorkActivity">
+    tools:context=".Vistas.Tareas.NewWorkActivity">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
