+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/Color2"
-    tools:context=".activities.Settings">
+    tools:context=".Vistas.Configuracion.Settings">
 
     <TextView
         android:id="@+id/textView11"
