+++ app/src/main/res/layout/activity_sync_up.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.SyncUp">
+    tools:context=".Vistas.Precarga.SyncUp">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline5"
