+++ app/src/main/res/layout/activity_works.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".task.WorksActivity">
+    tools:context=".Vistas.Tareas.WorksActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
